# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /transit/stops:
    get:
      operationId: getNearbyTransitStops
      summary: Nearby Krakow stops from official ZTP schedules, never demo data
      parameters:
        - { name: lat, in: query, required: true, schema: { type: number, minimum: -90, maximum: 90 } }
        - { name: lng, in: query, required: true, schema: { type: number, minimum: -180, maximum: 180 } }
        - { name: radius, in: query, schema: { type: integer, minimum: 100, maximum: 3000, default: 1500 } }
      responses:
        "200":
          description: Stops in the requested area
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NearbyTransitStops" }
        "400": { description: Invalid map position }
        "503": { description: Official timetable loading or unavailable }
  /transit/stops/{stopId}/departures:
    get:
      operationId: getTransitStopDepartures
      summary: Upcoming Krakow departures with fresh ZTP predictions or explicitly scheduled times
      parameters:
        - { name: stopId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Departures for the actual service date in Europe/Warsaw
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransitStopDepartures" }
        "404": { description: Unknown stop }
        "503": { description: Official timetable loading or unavailable }
...
  /maps/config:
    get:
      operationId: getGoogleMapsConfig
      summary: Restricted public browser map configuration (never the Routes API key)
      responses:
        "200":
          description: Browser map configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoogleMapsConfig"
        "503":
          description: Browser map key is missing
  /transit/routes:
    post:
      operationId: computeTransitRoutes
      summary: Real Google transit itineraries from device coordinates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitInput"
      responses:
        "200":
          description: Real transit routes, or an empty routes list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitResults"
        "400":
          description: Invalid coordinates
        "429":
          description: Request limit reached
        "503":
          description: Google configuration or service unavailable
  /geocoding/reverse:
    post:
      operationId: reverseAddress
      summary: Resolve real GPS coordinates using the shared geocoding capacity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitCoordinate"
      responses:
        "200":
          description: Sanitized GeoJSON address results, possibly empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressSearchResults"
        "400":
          description: Invalid coordinates
        "429":
          description: Shared geocoding capacity exhausted
        "503":
          description: Address resolution unavailable
  /geocoding/search:
    post:
      operationId: searchAddresses
      summary: Search addresses through the capacity-controlled Geoapify service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressSearchInput"
      responses:
        "200":
          description: Sanitized GeoJSON FeatureCollection (including empty results)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressSearchResults"
        "400":
          description: Invalid query
        "429":
          description: Shared capacity exhausted; Retry-After is provided
        "503":
          description: Provider not configured or temporarily unavailable
...
components:
  schemas:
    PassengerTransitStop:
      type: object
      required: [id, name, lat, lng, lines, transportTypes]
      properties:
        id: { type: string }
        name: { type: string }
        lat: { type: number }
        lng: { type: number }
        lines: { type: array, items: { type: string } }
        transportTypes: { type: array, items: { type: string, enum: [BUS, TRAM] } }
    PassengerDeparture:
      type: object
      required: [id, stopId, tripId, lineName, lineNumber, transportType, headsign, departureTime, scheduledDepartureTime, expectedDepartureTime, delaySeconds, status, realtimeUpdatedAt]
      properties:
        id: { type: string }
        stopId: { type: string }
        tripId: { type: string }
        lineName: { type: string }
        lineNumber: { type: string }
        transportType: { type: string, enum: [BUS, TRAM] }
        headsign: { type: string }
        departureTime: { type: string, format: date-time }
        scheduledDepartureTime:
          type: string
          format: date-time
          description: Planned time; departureTime remains the same for backwards compatibility.
        expectedDepartureTime: { type: ["string", "null"], format: date-time }
        delaySeconds: { type: ["integer", "null"] }
        status: { type: string, enum: [SCHEDULED, UPDATED, CANCELLED, SKIPPED] }
        realtimeUpdatedAt:
          type: ["string", "null"]
          format: date-time
          description: Source timestamp, not download time. Only predictions no older than 120 seconds are used.
    NearbyTransitStops:
      type: object
      required: [stops, generatedAt, cached, attribution]
      properties:
        stops: { type: array, items: { $ref: "#/components/schemas/PassengerTransitStop" } }
        generatedAt: { type: string, format: date-time }
        cached: { type: boolean }
        attribution: { type: string }
    TransitStopDepartures:
      type: object
      required: [stop, departures, generatedAt, cached, attribution, alerts, alertsStatus, alertsUpdatedAt]
      properties:
        stop: { $ref: "#/components/schemas/PassengerTransitStop" }
        departures: { type: array, items: { $ref: "#/components/schemas/PassengerDeparture" } }
        generatedAt: { type: string, format: date-time }
        cached: { type: boolean }
        attribution: { type: string }
        alerts: { type: array, items: { $ref: "#/components/schemas/PassengerServiceAlert" } }
        alertsStatus: { type: string, enum: [AVAILABLE, PARTIAL, UNAVAILABLE] }
        alertsUpdatedAt: { type: ["string", "null"], format: date-time }
    PassengerServiceAlert:
      type: object
      required: [id, title, description, effect, dataset, updatedAt, periods, appliesToStop, departureIds]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        effect: { type: integer }
        dataset: { type: string, enum: [A, M, T] }
        updatedAt: { type: string, format: date-time }
        periods: { type: array, items: { $ref: "#/components/schemas/PassengerAlertPeriod" } }
        appliesToStop: { type: boolean }
        departureIds: { type: array, items: { type: string } }
    PassengerAlertPeriod:
      type: object
      required: [start, end]
      properties:
        start: { type: ["string", "null"], format: date-time }
        end: { type: ["string", "null"], format: date-time }
...
    GoogleMapsConfig:
      type: object
      required: [apiKey]
      properties:
        apiKey: { type: string }
    TransitCoordinate:
      type: object
      additionalProperties: false
      required: [lat, lng]
      properties:
        lat: { type: number, minimum: -90, maximum: 90 }
        lng: { type: number, minimum: -180, maximum: 180 }
    TransitInput:
      type: object
      additionalProperties: false
      required: [origin, destination]
      properties:
        origin: { $ref: "#/components/schemas/TransitCoordinate" }
        destination: { $ref: "#/components/schemas/TransitCoordinate" }
    TransitAgency:
      type: object
      required: [name]
      properties:
        name: { type: string }
        uri: { type: string }
    TransitStep:
      type: object
      required: [mode, instruction, distanceMeters, durationSeconds]
      properties:
        mode: { type: string, enum: [WALK, TRANSIT, OTHER] }
        instruction: { type: string }
        distanceMeters: { type: ["number", "null"] }
        durationSeconds: { type: ["number", "null"] }
        polyline: { type: string }
        lineName: { type: string }
        lineShortName: { type: string }
        lineColor: { type: string }
        vehicleType: { type: string }
        headsign: { type: string }
        departureStop: { type: string }
        arrivalStop: { type: string }
        departureTime: { type: string }
        arrivalTime: { type: string }
        stopCount: { type: integer }
        agencies:
          type: array
          items: { $ref: "#/components/schemas/TransitAgency" }
    TransitRoute:
      type: object
      required: [id, distanceMeters, durationSeconds, polyline, transfers, steps]
      properties:
        id: { type: string }
        distanceMeters: { type: ["number", "null"] }
        durationSeconds: { type: ["number", "null"] }
        polyline: { type: string }
        transfers: { type: integer }
        fareText: { type: string }
        steps:
          type: array
          items: { $ref: "#/components/schemas/TransitStep" }
    TransitResults:
      type: object
      required: [routes, provider, requestedAt]
      properties:
        provider: { type: string, enum: [google] }
        requestedAt: { type: string }
        routes:
          type: array
          items: { $ref: "#/components/schemas/TransitRoute" }
    AddressSearchInput:
      type: object
      additionalProperties: false
      required: [query]
      properties:
        query:
          type: string
          minLength: 3
          maxLength: 200
    AddressSearchResults:
      type: object
      required: [features, provider]
      properties:
        provider:
          type: string
          enum: [geoapify]
        features:
          type: array
          maxItems: 8
          items:
            type: object
            additionalProperties: true
...