# lib/api-zod/src/generated/types/transitRoute.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransitStep } from './transitStep';

export interface TransitRoute {
  id: string;
  /** @nullable */
  distanceMeters: number | null;
  /** @nullable */
  durationSeconds: number | null;
  polyline: string;
  transfers: number;
  fareText?: string;
  steps: TransitStep[];
}